taskflow/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic[email]==2.7.1
pyjwt==2.8.0
bcrypt==4.1.3
python-multipart==0.0.9
pytest==8.2.0
httpx==0.27.0
python-dotenv==1.0.1
